+++ StatsArrList.java
@@ -1,11 +1,11 @@
 import java.util.ArrayList;
 public interface StatsArrList
 {
-    public int meanList(ArrayList<Integer> x);
+    public double meanList(ArrayList<Integer> x);
     
-    public int medianList(ArrayList<Integer> x);
+    public double medianList(ArrayList<Integer> x);
     
-    public int modeList(ArrayList<Integer> x);
+    public double modeList(ArrayList<Integer> x);
     
-    public int standardDeviationList(ArrayList<Integer> x);
+    public double standardDeviationList(ArrayList<Integer> x);
 }
